allow to parse multiple partitions and offsets
parse is now done in two steps.
During planing we create a list of KafkaScanOp.
This list can be red as rewritten kafka scan or conditions.
However this conditions don't necessarily have any upper bound
e.g. partition > 10

On Execution (after connection to kafka) we know the actual
available partitions and create a list of KafkaScanP which
is flattened (reduce overlaps from the former list) and contains
one (or more in case of offset gaps) entry for each partition to scan.