Add error handling at the end of kafka table scan.
When librdkafka establishes connection to server it starts few threads.
To properly release resources we should call kafkaCloseConnection(), but
unfortunately it never happend in case of error in kafkaEndForeignScan and
threads stayed active till the end of process. This commit wraps the code
which potentially can throw error with PG_TRY/PG_CATCH and properly ends
the connection.